extracted map handling into base map with first proxy and empty level1
