+++ braindefence/arcade/gamephase.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class GamePhase(Enum):
+    Running = (0,)
+    Won = (1,)
+    Lost = 2
