+++ braindefence/arcade/levels/__init__.py
@@ -1,0 +1,3 @@
+from braindefence.arcade.levels.base_map import BaseMap
+
+__all__ = ["BaseMap"]
