+++ braindefence/arcade/levels/level1map.py
@@ -1,3 +1,7 @@
+from braindefence.arcade.levels import BaseMap
 
-Class Level1Map(BaseMap) :
-l
+
+class Level1Map(BaseMap):
+
+    def __init__(self):
+        super(Level1Map, self).__init__(1)
